Autonomous Vehicle Refactor
