+++ Assets/UB_MR/Scripts/DigitalTwin/AutonomousVehicle.cs
@@ -5,87 +5,64 @@ using System.Collections.Generic;
 using vision_msgs.msg;
 using sensor_msgs.msg;
 using Unity.Cinemachine;
+using CAVAS.UB_MR.DT.VirtualObjectDetection;
 
 namespace CAVAS.UB_MR.DT
 {
-
     public class AutonomousVehicle : DigitalTwin
     {
         [Header("Cameras")]
         [SerializeField] CinemachineCamera dashCam;
         [SerializeField] CinemachineCamera followCam;
+
         [Space]
+
         [Header("Object Detection Settings")]
-        [SerializeField] bool enableVirtualObjectDetection = true; // Enable detection of virtual objects
-        [SerializeField] float virtualObjectDetectionRadius = 30f; // Radius in which virtual objects are detected
-        [SerializeField] string virtualObjectsTopicName = "/virtual_obstacles"; // Topic name for publishing virtual objects
-        [Space]
-        [Header("Image Capture Settings")]
+        [SerializeField] bool enableBoundingBoxCapture = true; // Enable detection of virtual objects
         [SerializeField] bool enableImageCapture = true; // Enable image capture
-        [SerializeField] Camera targetCamera; // Camera to capture images from
+        [SerializeField] string boundingBoxTopicName = "/virtual_obstacles"; // Topic name for publishing virtual object bounding boxes
+        [SerializeField] string virtualCameraTopicName = "/virtual_camera/image_raw/compressed"; // Topic name for publishing virtual camera images
+
+        [Space]
+
+        [Header("Bounding Box Parameters")]
+        [SerializeField] float detectionRadius = 30.0f; // Height of
+
+        [Header("Image Capture Parameters")]
         [SerializeField] int imageWidth = 640;
         [SerializeField] int imageHeight = 480;
         [SerializeField] float publishRate = 1.0f; // 1 FPS
-        [SerializeField] string frameId = "camera_link";
-        [SerializeField] string topicName = "/virtual_camera/image_raw/compressed";
 
-
-        ROS2Node mNode;
-        VirtualObjectDetector mVirtualObjectDetector;
-        ISubscription<nav_msgs.msg.Odometry> mWorldTransformationSubscriber;
-        IPublisher<BoundingBox3DArray> mObstacleBoundingBoxPublisher;
-        IPublisher<CompressedImage> imagePublisher;
-        RenderTexture renderTexture;
-        Texture2D texture2D;
 
         protected Vector3 mWorldPosition = Vector3.zero;
         protected Vector3 mAngularVelocity = Vector3.zero;
         protected Vector3 mLinearVelocity = Vector3.zero;
         protected Quaternion mWorldRotation = Quaternion.identity;
 
-        void ConnectToROS()
-        {
-            if (ROS2_Bridge.ROS_CORE.Ok() && this.mNode == null)
-            {
-                string name = gameObject.name.Replace("(Clone)", "");
-                name = name.Replace(" Variant", "");
-                // This is sort of cheating but ROS2_Bridge is not immediately deleting nodes so this avoids a collision
-                int randomSuffix = UnityEngine.Random.Range(0, 1000);
-                this.mNode = ROS2_Bridge.ROS_CORE.CreateNode(name + "_Digital_Twin_" + randomSuffix.ToString());
-                // World Transformation Subscriber
-                this.mWorldTransformationSubscriber = this.mNode.CreateSubscription<nav_msgs.msg.Odometry>("/world_transform", WorldTransformationUpdate);
-                // Obstacle Bounding Box Publisher
-                if (enableVirtualObjectDetection)
-                {
-                    this.mVirtualObjectDetector = new VirtualObjectDetector(this.transform);
-                    this.mObstacleBoundingBoxPublisher = this.mNode.CreatePublisher<BoundingBox3DArray>(virtualObjectsTopicName);
-                }
+        ROS2Node mNode;
+        VirtualBoundingBoxDetector mVirtualBoundingBoxDetector;
+        VirtualCameraOverlay mVirtualCameraOverlay;
+        ISubscription<nav_msgs.msg.Odometry> mWorldTransformationSubscriber;
 
-                // Image Publisher
-                if (enableImageCapture)
-                {
-                    targetCamera = FindFirstObjectByType<Camera>();
-                    imagePublisher = this.mNode.CreatePublisher<CompressedImage>(topicName);
-                    // Create render texture and texture2D for image capture
-                    renderTexture = new RenderTexture(imageWidth, imageHeight, 24);
-                    texture2D = new Texture2D(imageWidth, imageHeight, TextureFormat.RGB24, false);
-                    StartCoroutine(PublishActiveCameraImage());
-                }
-            }
-        }
+        
 
         public override void OnNetworkSpawn()
         {
             base.OnNetworkSpawn();
             if (IsOwner)
-            {
                 ConnectToROS();
-            }
+            StartCoroutine(PublishVirtualObjects());
         }
 
         public override void OnDestroy()
         {
             base.OnDestroy();
+            if (IsOwner)
+            {
+                this.mVirtualCameraOverlay.CleanUp();
+                this.mVirtualBoundingBoxDetector.CleanUp();
+            }
+            
 
             if (this.mNode != null && this.mWorldTransformationSubscriber != null)
             {
@@ -99,68 +76,38 @@ namespace CAVAS.UB_MR.DT
                 ROS2_Bridge.ROS_CORE.RemoveNode(this.mNode);
                 this.mNode = null;
             }
+        }
 
-            if (this.mNode != null)
+        void ConnectToROS()
+        {
+            if (ROS2_Bridge.ROS_CORE.Ok() && this.mNode == null)
             {
-                // Clean up resources
-                if (renderTexture != null)
-                    renderTexture.Release();
-                if (texture2D != null)
-                    Destroy(texture2D);
+                string name = gameObject.name.Replace("(Clone)", "");
+                name = name.Replace(" Variant", "");
+                // This is sort of cheating but ROS2_Bridge is not immediately deleting nodes so this avoids a collision
+                int randomSuffix = UnityEngine.Random.Range(0, 1000);
+                this.mNode = ROS2_Bridge.ROS_CORE.CreateNode(name + "_Digital_Twin_" + randomSuffix.ToString());
+                // World Transformation Subscriber
+                this.mWorldTransformationSubscriber = this.mNode.CreateSubscription<nav_msgs.msg.Odometry>("/world_transform", WorldTransformationUpdate);
+                // Obstacle Bounding Box Publisher
+                if (enableBoundingBoxCapture)
+                {
+                    this.mVirtualBoundingBoxDetector = new VirtualBoundingBoxDetector(boundingBoxTopicName, this.mNode, this.transform);
+                    this.mVirtualCameraOverlay = new VirtualCameraOverlay(virtualCameraTopicName, this.mNode, FindFirstObjectByType<Camera>());
+                }
             }
         }
 
-        void Update()
+        public IEnumerator PublishVirtualObjects()
         {
-            if (IsOwner && enableVirtualObjectDetection)
+            while (IsOwner)
             {
-                PublishNearbyVirtualObjects();
+                if (enableBoundingBoxCapture)
+                    this.mVirtualBoundingBoxDetector.PublishNearbyVirtualObjects(detectionRadius);
+                if (enableImageCapture)
+                    this.mVirtualCameraOverlay.CaptureAndPublishImage(imageWidth, imageHeight);
+                yield return new WaitForSeconds(1.0f / publishRate);
             }
-        }
-
-        void PublishNearbyVirtualObjects()
-        {
-            // TODO: Bounding Boxes position and orientation should be relative to the Ego-Vehicle, not the world
-            vision_msgs.msg.BoundingBox3DArray msg = new vision_msgs.msg.BoundingBox3DArray();
-            // Header
-            msg.Header = new std_msgs.msg.Header();
-            msg.Header.Frame_id = "world";
-            builtin_interfaces.msg.Time time = new builtin_interfaces.msg.Time();
-            time.Sec = (int)UnityEngine.Time.timeSinceLevelLoad;
-            msg.Header.Stamp = time;
-
-            List<VirtualObject> virtualObjects = this.mVirtualObjectDetector.GetNearbyObstacles(virtualObjectDetectionRadius);
-            var boxes = new BoundingBox3D[virtualObjects.Count];
-            for (int i = 0; i < boxes.Length; i++)
-            {
-                VirtualObject virtualObject = virtualObjects[i];
-                Bounds bounds = virtualObject.GetBoundingBox();
-                BoundingBox3D bbox = new BoundingBox3D();
-                bbox.Center = new geometry_msgs.msg.Pose();
-                // Position
-                bbox.Center.Position = new geometry_msgs.msg.Point();
-                bbox.Center.Position.X = bounds.center.x;
-                bbox.Center.Position.Y = bounds.center.y;
-                bbox.Center.Position.Z = bounds.center.z;
-                // Orientation
-                bbox.Center.Orientation = new geometry_msgs.msg.Quaternion();
-                bbox.Center.Orientation.X = virtualObject.transform.rotation.x;
-                bbox.Center.Orientation.Y = virtualObject.transform.rotation.y;
-                bbox.Center.Orientation.Z = virtualObject.transform.rotation.z;
-                bbox.Center.Orientation.W = virtualObject.transform.rotation.w;
-                // Size
-                bbox.Size = new geometry_msgs.msg.Vector3();
-                bbox.Size.X = bounds.size.x;
-                bbox.Size.Y = bounds.size.y;
-                bbox.Size.Z = bounds.size.z;
-
-                boxes[i] = bbox;
-            }
-            msg.Boxes = boxes;
-            msg.WriteNativeMessage();
-            this.mObstacleBoundingBoxPublisher.Publish(msg);
-
-            //Debug.Log("Published " + virtualObjects.Count + " virtual objects to topic: " + virtualObjectsTopicName);
         }
 
         public virtual Vector3 GetLinearVelocity()
@@ -171,6 +118,49 @@ namespace CAVAS.UB_MR.DT
         public virtual Vector3 GetAngularVelocity()
         {
             return this.mAngularVelocity;
+        }
+
+        public void EnableDashCam(bool inEnable)
+        {
+            if (IsOwner)
+            {
+                base.EnableCameras(false);
+                if (dashCam != null)
+                    dashCam.gameObject.SetActive(inEnable);
+            }
+        }
+
+        public void EnableFollowCam(bool inEnable)
+        {
+            if (IsOwner)
+            {
+                base.EnableCameras(false);
+                if (followCam != null)
+                    followCam.gameObject.SetActive(inEnable);
+            }
+        }
+
+        public void SetLayerCulling(Camera camera, string layerName, bool shouldRender)
+        {
+            int layerIndex = LayerMask.NameToLayer(layerName);
+            if (layerIndex == -1)
+            {
+                Debug.LogWarning($"Layer '{layerName}' does not exist.");
+                return;
+            }
+
+            int layerMask = 1 << layerIndex;
+            if (shouldRender)
+            {
+                camera.cullingMask |= layerMask;
+                camera.clearFlags = CameraClearFlags.Skybox;
+            }
+            else
+            {
+                camera.cullingMask &= ~layerMask;
+                camera.clearFlags = CameraClearFlags.SolidColor;
+                camera.backgroundColor = Color.black;
+            }
         }
 
         void WorldTransformationUpdate(nav_msgs.msg.Odometry msg)
@@ -210,86 +200,7 @@ namespace CAVAS.UB_MR.DT
             );
         }
 
-        void CaptureAndPublishImage()
-        {
-            if (imagePublisher == null || targetCamera == null)
-                return;
-                
-            // Capture camera image (your existing code)
-            RenderTexture currentRT = RenderTexture.active;
-            targetCamera.targetTexture = renderTexture;
-            targetCamera.Render();
-            RenderTexture.active = renderTexture;
-            texture2D.ReadPixels(new Rect(0, 0, imageWidth, imageHeight), 0, 0);
-            texture2D.Apply();
-            
-            // Restore render texture
-            targetCamera.targetTexture = null;
-            RenderTexture.active = currentRT;
-            
-            // Convert to JPEG and publish 
-            byte[] imageBytes = texture2D.EncodeToJPG(75);
-            var compressedImage = new CompressedImage();
-            builtin_interfaces.msg.Time time = new builtin_interfaces.msg.Time();
-            time.Sec = (int)UnityEngine.Time.timeSinceLevelLoad; // Use Time.timeSinceLevelLoad for simulation time
-            compressedImage.Header.Stamp = time;
-            compressedImage.Header.Frame_id = frameId;
-            compressedImage.Format = "jpeg";
-            compressedImage.Data = imageBytes;
-            imagePublisher.Publish(compressedImage);
-        }
 
-        public void EnableDashCam(bool inEnable)
-        {
-            if (IsOwner)
-            {
-                base.EnableCameras(false);
-                if (dashCam != null)
-                    dashCam.gameObject.SetActive(inEnable);
-            }
-        }
-
-        public void EnableFollowCam(bool inEnable)
-        {
-            if (IsOwner)
-            {
-                base.EnableCameras(false);
-                if (followCam != null)
-                    followCam.gameObject.SetActive(inEnable);
-            }
-        }
-
-        public void SetLayerCulling(Camera camera, string layerName, bool shouldRender)
-        {
-            int layerIndex = LayerMask.NameToLayer(layerName);
-            if (layerIndex == -1)
-            {
-                Debug.LogWarning($"Layer '{layerName}' does not exist.");
-                return;
-            }
-            
-            int layerMask = 1 << layerIndex;
-            if (shouldRender)
-            {
-                camera.cullingMask |= layerMask;
-                camera.clearFlags = CameraClearFlags.Skybox;
-            }
-            else
-            {
-                camera.cullingMask &= ~layerMask;
-                camera.clearFlags = CameraClearFlags.SolidColor;
-                camera.backgroundColor = Color.black;
-            }
-        }
-
-        IEnumerator PublishActiveCameraImage()
-        {
-            while (true)
-            {
-                CaptureAndPublishImage();
-                yield return new WaitForSeconds(1.0f / publishRate);
-            }
-        }
     }
 
 }
