+++ Assets/UB_MR/Scripts/DigitalTwin/VirtualObjectDetection.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f4d6242735a7689959e5eddbb65e832a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
