+++ Assets/UB_MR/Scripts/DigitalTwin/VirtualObjectDetection/VirtualCameraOverlay.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a5bc0aa8de21c46ae8175201d4df0738
